♻️ certif:  fusion acceptance tests
